feat: implement financial dashboard data access objects and models
