+++ Models/FinancialDashboard/PivDivisionModel.cs
@@ -1,0 +1,9 @@
+namespace MISReports_Api.Models.FinancialDashboard
+{
+    public class PivDivisionModel
+    {
+        public string date { get; set; }
+        public string company { get; set; }
+        public double amount { get; set; }
+    }
+}
