+++ Models/FinancialDashboard/PivTotalModel.cs
@@ -1,0 +1,8 @@
+namespace MISReports_Api.Models.FinancialDashboard
+{
+    public class PivTotalModel
+    {
+        public string date { get; set; }
+        public double amount { get; set; }
+    }
+}
