+++ Models/FinancialDashboard/StockDivisionModel.cs
@@ -1,0 +1,8 @@
+namespace MISReports_Api.Models.FinancialDashboard
+{
+    public class StockDivisionModel
+    {
+        public string company { get; set; }
+        public double amount { get; set; }
+    }
+}
